### 2023-07-29 #5707 @danielskatz
Additionally, in the .md file, multiple references should be put together as
`[@rybakken-et-al; @kang-xu-morozov; @gardner-et-al]`
and a reference that is a direct mention in the paper doesn't need the `[]`s around it.  Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
