### 2023-07-31 e9fd69c
add missing DOIs
