+++ paper/paper.bib
@@ -126,7 +126,8 @@ MRREVIEWER = {Patrizio\ Frosini},
   number={1},
   pages={5--26},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s00200-015-0251-x}
 }
 
 @InProceedings{fibered,
@@ -155,7 +156,8 @@ MRREVIEWER = {Patrizio\ Frosini},
   volume={15},
   pages={616748},
   year={2021},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/fncom.2021.616748}
 }
 
 @article{gardner-et-al,
@@ -166,7 +168,8 @@ MRREVIEWER = {Patrizio\ Frosini},
   number={7895},
   pages={123--128},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-021-04268-7}
 }
 
 @article{rybakken-et-al,
@@ -177,5 +180,6 @@ MRREVIEWER = {Patrizio\ Frosini},
   number={1},
   pages={68--93},
   year={2019},
-  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209, USA journals-info~…}
+  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209, USA journals-info~…},
+  doi={10.1162/neco_a_01150}
 }
