+++ paper/paper.md
@@ -31,10 +31,11 @@ Currently, DREiMac is able to find topology-preserving representations of point 
 In a few words, DREiMac takes as input a point cloud together with a topological feature of the point cloud (in the form of a persistent cohomology class), and returns a map from the point cloud to a well-understood topological space (a circle, a product of circles, a projective space, or a lens space), which preserves the given topological feature in a precise sense.
 
 DREiMac is based on persistent cohomology [@persistent-cohomology], a method from topological data analysis; the theory behind DREiMac is developed in
-[@circular-coords],
-[@projective-coords],
-[@lens-coords],
-[@toroidal-coords].
+@circular-coords,
+@projective-coords,
+@lens-coords,
+and
+@toroidal-coords.
 DREiMac is implemented in Python, using Numba for the more expensive computations.
 We test DREiMac periodically in Ubuntu, macOS, and Windows.
 
@@ -44,13 +45,13 @@ The documentation for DREiMac can be found [here](https://scikit-tda.org/DREiMac
 # Statement of need and main contributions
 
 Topological coordinatization is witnessing increased application in domains such as
-neuroscience [@rybakken-et-al] [@kang-xu-morozov] [@gardner-et-al],
+neuroscience [@rybakken-et-al; @kang-xu-morozov; @gardner-et-al],
 dynamical systems [@vejdemo-pokorny-skraba-kragic],
 and dimensionality reduction [@fibered].
 The fast implementations and data science integrations provided in DREiMac are aimed at enabling other domain scientists in their pursuits.
 
 To the best of our knowledge, the only publicly available software implementing cohomological coordinates based on persistent cohomology is Dionysus [@dionysus].
-Dionysus is a general purpose library for topological data analysis, which in particular implements the original circular coordinates algorithm of [@desilva-morozov-vejdemo].
+Dionysus is a general purpose library for topological data analysis, which in particular implements the original circular coordinates algorithm of @desilva-morozov-vejdemo.
 
 DREiMac adds to the current landscape of cohomological coordinates software by implementing various currently missing functionalities; we elaborate on these below.
 DREiMac also includes functions for generating topologically interesting datasets for testing, various geometrical utilities including functions for manipulating the coordinates returned by the algorithms, and several example notebooks including notebooks illustrating the effect of each of the main parameters of the algorithms.
@@ -59,10 +60,7 @@ DREiMac also includes functions for generating topologically interesting dataset
 All of DREiMac's coordinates are sparse, meaning that persistent cohomology computations are carried on a simplicial complex built on a small sample of the full point cloud.
 This gives a significant speedup when compared to algorithms which use a simplicial complex built on the entire dataset, since the persistent cohomology computation is the most computationally intensive part of the algorithm.
 For a precise description of the notion of sparseness, see the papers that develop the algorithms that DREiMac implements 
-[@circular-coords],
-[@projective-coords],
-[@lens-coords],
-[@toroidal-coords].
+[@circular-coords; @projective-coords; @lens-coords; @toroidal-coords].
 
 
 **Improvements to the circular coordinates algorithm.**
@@ -71,7 +69,7 @@ DREiMac implements two new functionalities addressing two issues that can arise 
 The circular coordinates algorithm turns a cohomology class with coefficients in $\mathbb{Z}$ into a map into the circle.
 However, since persistent cohomology is computed with coefficients in a field, the cohomology class is obtained by lifting a cohomology class with coefficients in $\mathbb{Z}/q\mathbb{Z}$, with $q$ a prime.
[…]
