+++ paper/paper.bib
@@ -83,8 +83,8 @@ MRREVIEWER = {Ne\v{z}a\ Mramor Kosta},
                   Jean{-}Lou De Carufel},
   title        = {Coordinatizing Data With Lens Spaces and Persistent Cohomology},
   booktitle    = {Proceedings of the 31st Canadian Conference on Computational Geometry,
-                  {CCCG} 2019, August 8-10, 2019, University of Alberta, Edmonton, Alberta,
-                  Canada},
+                  {CCCG} 2019, {A}ugust 8-10, 2019, {U}niversity of {A}lberta, {E}dmonton, {A}lberta,
+                  {C}anada},
   pages        = {49--58},
   year         = {2019},
   timestamp    = {Thu, 23 Apr 2020 16:22:52 +0200},
