### 2023-08-21 #5707 @LuisScoccola
Potential reviewers we are aware of:
- From the reviewers already registered in JOSS: _EduPH_, _yossibokor_, _ismailguzel_, _leotrs_, _rrrlw_, _corybrunson_
- Contributors to [rivet](https://github.com/rivetTDA/rivet), such as _xoltar_, _mlwright84_
- Contributors to [giotto-tda](https://giotto-ai.github.io/gtda-docs/0.5.1/library.html) or [giotto-deep](https://joss.theoj.org/papers/10.21105/joss.04846), such as _ulupo_, _gtauzin_, _matteocao_, _raphaelreinauer_
