### 2023-08-23 #5707 @fabian-s
👋 hi @leotrs @mtsch  @lewtun

would any of you be willing to review this submission for JOSS?
