### 2023-08-30 #5707 @fabian-s
👋 hi @lewtun @gtauzin  @raphaelreinauer

would any of you be willing to review this submission for JOSS?
