### 2023-08-31 #5791 @editorialbot
[REVIEW]: DREiMac: Dimensionality Reduction with Eilenberg-MacLane Coordinates
**Submitting author:** @LuisScoccola (<a href="http://orcid.org/0000-0002-4862-722X">Luis Scoccola</a>)
**Repository:** https://github.com/scikit-tda/DREiMac
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.3.0
