### 2023-10-03 scikit-tda/DREiMac#11 @mtsch
[JOSS] Software references in paper
Hi, first of all thanks for the well-written paper and great software! I love the detailed documentation with lots of examples, and how intuitive and easy to use it is.

I have a small comment regarding citing other software that you build upon. In particular, I think you should mention Ripser and Ripser.py as those are crucial to quickly computing persistent cohomology used to find circular coordinates. It might also be worth mentioning other software you use that has papers written on them (I could find papers on numba, matplotlib and numpy, you can decide which of these you want to include).

On a related note, my software package Ripserer also implements (a more limited form of) sparse circular coordinates. I understand that's easy to miss as I never advertised it and it is somewhat hidden in the [docs](https://mtsch.github.io/Ripserer.jl/dev/api/#Ripserer.CircularCoordinates). You could mention it alongside Dionysus in the statement of need.

Ref. [#5791](https://github.com/openjournals/joss-reviews/issues/5791)
