### 2023-10-03 scikit-tda/DREiMac#11 @ctralie
Thanks for the feedback!  Very interesting that you had your own
implementation of sparse circular coordinates!  We can certainly link to
that from our documentation
