### 2023-10-03 f2c6a88
add further references
