+++ paper/paper.bib
@@ -183,3 +183,53 @@ MRREVIEWER = {Patrizio\ Frosini},
   publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209, USA journals-info~…},
   doi={10.1162/neco_a_01150}
 }
+
+@article{ripserer,
+  doi = {10.21105/joss.02614},
+  url = {https://doi.org/10.21105/joss.02614},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {54},
+  pages = {2614},
+  author = {Matija Čufar},
+  title = {Ripserer.jl: flexible and efficient persistent homology computation in Julia},
+  journal = {Journal of Open Source Software}
+}
+
+@inproceedings{numba,
+  author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+  title = {Numba: A LLVM-Based Python JIT Compiler},
+  year = {2015},
+  isbn = {9781450340052},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/2833157.2833162},
+  doi = {10.1145/2833157.2833162},
+  abstract = {Dynamic, interpreted languages, like Python, are attractive for domain-experts and scientists experimenting with new ideas. However, the performance of the interpreter is often a barrier when scaling to larger data sets. This paper presents a just-in-time compiler for Python that focuses in scientific and array-oriented computing. Starting with the simple syntax of Python, Numba compiles a subset of the language into efficient machine code that is comparable in performance to a traditional compiled language. In addition, we share our experience in building a JIT compiler using LLVM[1].},
+  booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+  articleno = {7},
+  numpages = {6},
+  keywords = {Python, LLVM, compiler},
+  location = {Austin, Texas},
+  series = {LLVM '15}
+}
+
+@article{ripserpy, doi = {10.21105/joss.00925}, url = {https://doi.org/10.21105/joss.00925}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {29}, pages = {925}, author = {Christopher Tralie and Nathaniel Saul and Rann Bar-On}, title = {Ripser.py: A Lean Persistent Homology Library for Python}, journal = {Journal of Open Source Software} }
+
+@article {ripser,
+    AUTHOR = {Bauer, Ulrich},
+     TITLE = {Ripser: efficient computation of {V}ietoris-{R}ips persistence
+              barcodes},
+   JOURNAL = {J. Appl. Comput. Topol.},
+  FJOURNAL = {Journal of Applied and Computational Topology},
+    VOLUME = {5},
+      YEAR = {2021},
+    NUMBER = {3},
+     PAGES = {391--423},
+      ISSN = {2367-1726,2367-1734},
+   MRCLASS = {55N31 (55-04)},
+  MRNUMBER = {4298669},
+       DOI = {10.1007/s41468-021-00071-5},
+       URL = {https://doi.org/10.1007/s41468-021-00071-5},
+}
