### 2023-10-03 scikit-tda/DREiMac#11 @LuisScoccola
Hi, thank you for the feedback! Please see the latest version of the paper and let us know if we have missed anything.
