### 2023-10-04 b3f3aec
[JOSS] Update paper.md
I figured it's easier to fix with PR, since it's just a small typo/weird wording.
