### 2023-10-04 scikit-tda/DREiMac#11 @mtsch
Thanks, looks good to me now!
