### 2023-10-04 #5791 @mtsch
@fabian-s I've completed my review. The paper is well-written and covers everything without being too long, the documentation is detailed and full of examples, and the software works exactly as described. I think DREiMac will make a great entry in JOSS!
