### 2023-10-05 #5791 @fabian-s
@raphaelreinauer -- your checklist remains unfilled, is this due to issues with the paper/software you have identified?
if so, please comment and ask questions on this thread re. the open items on your checklists to help @LuisScoccola resolve any issues you have come across.

If not, please check the items that are in order so we can proceed with the review process.

Note that the JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues (or pull requests, even) on the software repository. When doing so, please mention openjournals/joss-reviews#5791 so that a link is created to this thread (and I can keep an eye on what is happening).

Thank you!
