### 2023-10-05 #5791 @fabian-s
For more complicated matters of discussion or code issues, reviewers are encouraged to submit issues (or pull requests, even) on the software repository for the paper linked at the top of this thread. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/5791 so that a link is created to this thread (and I can keep an eye on what is happening).

For simpler matters, simply asking in this thread is fine as well.
