### 2023-10-05 #5791 @raphaelreinauer
One of the contributors [Thomas Mease](https://github.com/Tmease5580), appears as someone other than the author of the submitted paper. Additionally, the authors' order does not reflect the number of contributions to the repository. If the order is alphabetical, I think it would be helpful to clarify in a section the individual contributions.
@LuisScoccola, Could you please clarify?
