### 2023-10-06 #5791 @LuisScoccola
Hi @raphaelreinauer . Thank you for your feedback!

Order is indeed alphabetical; please see the "Statement of contribution" section that I just added to the paper. With respect to the other question, Thomas contributed code that is not currently in the library and is thus not part of the submission. Please let us know if this further clarification is needed!
