### 2023-10-08 #5791 @raphaelreinauer
@LuisScoccola, I took a closer look at the repository, and it appears there are indeed commits from https://github.com/Tmease5580 in the paper branch, specifically [this commit](https://github.com/scikit-tda/DREiMac/commit/78ab62c2b49288c3908b2f690828dc9d8f9397b1). This suggests that Thomas Mease's contributions are indeed part of the paper submission. Could you please clarify this discrepancy?
