### 2023-10-08 #5791 @LuisScoccola
Hi @raphaelreinauer. To clarify further: Thomas’ commits to master can be found [here](https://github.com/scikit-tda/DREiMac/commits?author=Tmease5580). Said commits correspond to experimental features in an alternative javascript version, which did not get off the ground - please note that the files modified in these commits are not part of the current codebase. Since the codebase being submitted does not build on this code, either directly by using it or indirectly by reimplementing it in Python, it is our opinion that these commits alone do not meet the threshold of contribution needed for inclusion in this submission.
