### 2023-10-10 #5791 @raphaelreinauer
Thank you for the detailed clarification, @LuisScoccola. I took the initiative to check manually and can confirm that the files edited by Thomas in those commits are indeed deleted in the current head of the main branch. It's evident that those contributions are not part of the codebase being submitted.

As a best practice, experimental features, especially ones that are not intended for the final version, should ideally not be merged into the main branch.

For clarity and to avoid future confusion, you might consider purging those files from the repository's history as described [here](https://www.30secondsofcode.org/git/s/purge-file/).
