### 2023-10-10 #5791 @raphaelreinauer
@LuisScoccola
While reviewing the repository, I went through some of the notebooks and noticed that a couple of them threw exceptions:

1. In `parameter_standard_range.ipynb`, I encountered the following error:
```
Exception                                 Traceback (most recent call last)
DREiMac/docs/notebooks/parameter_standard_range.ipynb Cell 8 line 1
----> 1 circular_coordinates = cc2.get_coordinates(standard_range=True)
```
2. Similarly, in `complexProjectiveOnGenusTwoSurface.ipynb`, I stumbled upon:
```
DREiMac/docs/notebooks/complexProjectiveOnGenusTwoSurface.ipynb Cell 7 line 1
----> 1 get_ipython().run_line_magic('matplotlib', 'widget')
      2 fig = plt.figure(figsize=(8,4))
      3 ax = fig.add_subplot(projection="3d")

ModuleNotFoundError: No module named 'ipympl'
```
For this particular issue, it might be a good idea to add `ipympl `to the example requirements to avoid any dependency issues.

Furthermore, as a proactive measure, I'd recommend setting up a GitHub action that automatically checks if all Jupyter notebooks run without exception for each PR to the main branch.
