### 2023-10-10 #5791 @raphaelreinauer
https://github.com/scikit-tda/DREiMac/blob/9ced70db1f275c9fcaa78c1788477e764b1106f9/dreimac/toroidalcoords.py#L140-L146
This could be simplified to:
```
delta1 = delta1 or CohomologyUtils.make_delta1_compact(
    dist_land_land, rips_threshold, self.cns_lookup_table_
)
```
