### 2023-10-10 #5791 @raphaelreinauer
The naming of protected variables in the code, like `self.prime_`, `self.dgms_`, and `self.idx_land_` in `EMCoords`, doesn't follow Python conventions. The convention is to prefix them with a single underscore (e.g., self._prime).
