### 2023-10-10 #5791 @raphaelreinauer
The variable `self.type_` in the classes `EMCoords`, `ToroidalCoords`, and `ComplexProjectiveCoords`, ...  is seemingly used to indicate the type.

This seems redundant as the instance type can be determined with the `type()` function. Consider removing it if it's not utilized elsewhere.
