### 2023-10-17 0be2c4f
fix: add <3.12 constraint to python version
