+++ README.md
@@ -58,7 +58,7 @@ For Jupyter notebooks with more examples, please check the [documentation](https
 
 ## Installing
 
-Make sure you are using Python 3.8 or newer.
+Make sure your Python version is >=3.8 and <3.12.
 DREiMac depends on the following python packages, which will be installed automatically when you install with pip:
 `matplotlib`,
 `numba`,
