### 2023-10-17 8902a4f
style: replace if with or
