### 2023-10-17 6ff9640
style: delete unused type attributes
