### 2023-10-17 a18ff0e
style: delete unused type attributes
