### 2023-10-17 ba53605
style: use leading underscore convention for private attributes
