### 2023-10-17 01e77d7
style: use leading underscore convention for private attributes
