### 2023-10-17 #5791 @LuisScoccola
@raphaelreinauer thank you very much for your suggestions!

Please see the [pull request](https://github.com/scikit-tda/DREiMac/pull/13) implementing most of your suggestions and let us know what you think. Let me comment on the suggestions that have not been addressed:
- We haven't purged any files from the history since, as far as I understand, this cannot be undone, so we would like to give it a bit more thought.
- The error in `parameter_standard_range.ipynb` is intentional; we are showcasing what happens when an invalid selection of parameters is done. We are happy to hear suggestions on how to handle this, however.
- The utils have not been refactored into their own files. Some of the future developments we have in mind would require to have some internal attributes, so, although it is not clear when these will be implemented, we believe it is simpler to keep it this way for now.
