+++ dreimac/toroidalcoords.py
@@ -43,6 +43,7 @@ class ToroidalCoords(EMCoords):
 
     def get_coordinates(
         self,
+        X_query=None,
         perc=0.5,
         cocycle_idxs=[0],
         partunity_fn=PartUnity.linear,
@@ -54,6 +55,8 @@ class ToroidalCoords(EMCoords):
 
         Parameters
         ----------
+        X_query: ndarray(M, d) or None
+            A point cloud to compute the toroidal coordinates on. If None, uses self.X.
         perc : float
             Percent coverage. Must be between 0 and 1.
         cocycle_idx : integer
@@ -71,11 +74,13 @@ class ToroidalCoords(EMCoords):
 
         Returns
         -------
-        thetas : ndarray(n, N)
+        thetas : ndarray(n, M)
             List of circular coordinates, with n the length of cocycle_idxs
 
         """
-
+        if (X_query is not None) and self.distance_matrix:
+            raise NotImplementedError("To compute coordinates on points other than the input, "
+                                      "initialize ToroidalCoords with points instead of a distance matrix.")
         # get representative cocycles and the intersection of their supports
         homological_dimension = 1
         cohomdeaths, cohombirths, cocycles = zip(
@@ -106,7 +111,7 @@ class ToroidalCoords(EMCoords):
         )
 
         # compute partition of unity and choose a cover element for each data point
-        varphi, ball_indx = EMCoords.get_covering_partition(self, r_cover, partunity_fn)
+        varphi, ball_indx = EMCoords.get_covering_partition(self, r_cover, partunity_fn, X_query)
 
         # compute boundary matrix
         dist_land_land = self._dist_land_land
