### 2023-10-17 71ffea9
fix: propagate change in naming convention of private attribute
