### 2023-10-18 7d638d2
fix: keep old naming convention for internal variable to not break interface until next interface-breaking version
