### 2023-10-19 #5791 @raphaelreinauer
> Similarly, in complexProjectiveOnGenusTwoSurface.ipynb, I stumbled upon:
> ModuleNotFoundError                       Traceback (most recent call last)
> DREiMac/docs/notebooks/complexProjectiveOnGenusTwoSurface.ipynb Cell 7 line 1
> ----> 1 get_ipython().run_line_magic('matplotlib', 'widget')
>       2 fig = plt.figure(figsize=(8,4))
>       3 ax = fig.add_subplot(projection="3d")
>
> ModuleNotFoundError: No module named 'ipympl'
> Here ipympl is missing in the example requirements.

I couldn't find any comments/changes from you on this.
