### 2023-10-19 #5791 @raphaelreinauer
> We haven't purged any files from the history since, as far as I understand, this cannot be undone, so we would like to give it a bit more  thought.

A suggestion: consider making a local clone of the repository, testing the file purge there, and if successful, then pushing the changes to GitHub.
