### 2023-10-19 #5791 @raphaelreinauer
> The utils have not been refactored into their own files. Some of the future developments we have in mind would require to have some internal attributes, so, although it is not clear when these will be implemented, we believe it is simpler to keep it this way for now.

I understand the foresight for future developments, but I still need to see the current necessity of classes. If the need arises, it's better to reevaluate the design rather than keep it as is now.
