### 2023-10-19 #5791 @raphaelreinauer
> Furthermore, as a proactive measure, I'd recommend setting up a GitHub action that automatically checks if all Jupyter notebooks run without exception for each PR to the main branch.

Could you please comment on this?
