### 2023-10-20 #5791 @fabian-s
@raphaelreinauer thanks a lot for the dilligent code review, JOSS sincerely appreciates the effort.

Could you please update your checklist above and check off the items that have been resolved to your satisfaction?

Note that it's important for peer reviewer to distinguish between a) actual bugs, standard violations and clearly unfilled checklist items and b) differences in coding style & taste or personal preferences for specific workflows.

IMO as action editor,

- decisions on whether/how to clean up a repo's commit history or branches
- details of how the repo under review decides to set up their CI and PR hooks
- whether working and reasonably well-documented, but potentially slightly suboptimal code should be refactored immediately or left as is

all fall into category b), are up to the authors themselves and I would not consider such issues as blocking acceptance of a submission.
Once again, please take this in the spirit as intended: not as a rebuke or dismissal of your dilligent review work, but as a clarification from my side what I consider to be in and out of scope for a JOSS review, and a request  for clarification from you about which actually blocking issues in the checklist remain open.
