### 2023-10-20 #5791 @raphaelreinauer
I couldn't find any benchmarks in the repository verifying this claim. Could you please show me where to look?

> This gives a significant speedup when compared to algorithms which use a simplicial complex built on the entire dataset, since the persistent cohomology computation is the most computationally intensive part of the algorithm.

 (L40-L42)
