### 2023-10-20 #5791 @raphaelreinauer
@fabian-s Thanks for the clarification on the scope of the review.
I've updated the checklist as you asked.

This question is under the point "Installation" in the checklist:
https://github.com/openjournals/joss-reviews/issues/5791#issuecomment-1771647927

Automatically testing the functionality of the notebooks falls under the point " Automated tests".

All other feedback aligns with category b) and is more suggestions than imperative changes.

In future comments, I will reference the checklist to clarify where they fit in the review framework. This should help distinguish between essential and optional feedback, making the process smoother for all involved.
