### 2023-10-22 #5791 @raphaelreinauer
@fabian-s

Concerning the appearance of Thomas Mease (Tmease5580) as an author of the repository, could you please give me some guidance or your decision on whether and how we should address cleaning up the repository's commit history or branches in this context?

![image](https://github.com/openjournals/joss-reviews/assets/16176161/e00dce50-1501-4e3c-ba09-4ccff0cee65d)
