### 2023-10-23 #5791 @fabian-s
>Thanks for the clarification on the scope of the review.
> I've updated the checklist as you asked.

Thank **you** for your dilligent and constructive review!

> Concerning the appearance of Thomas Mease (Tmease5580) as an author of the repository (about the point "Contribution and authorship" in the checklist), could you please give me some guidance or your decision on whether and how we should address cleaning up the repository's commit history or branches in this context?

Since the code contributed by TM is not part of the code that is described in the submission, it's clearly the right call to not include them as an author of the paper.
That his commits remain part of the revision history on main is not a big deal IMO, just slightly messy, and whether to purge them or not is up to the team working on the repo, IMO.
we try not to interfere with people's workflows or coding style as much as possible as long as the submitted software submitted does what it says and the submitted paper says what it does.
