### 2023-10-24 9adccdc
test: check that notebooks run without errors, catch exception in a notebook
