+++ .github/workflows/tests.yml
@@ -25,10 +25,10 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install pytest
+          pip install pytest nbmake
       - name: Run test 
         run: |
-          pytest .
+          pytest --nbmake .
 
 
   test-windows:
@@ -48,11 +48,11 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install -r requirements.txt
-          pip install pytest
+          pip install pytest nbmake
       - name: Run test 
         shell: bash -l {0}
         run: |
-          pytest .
+          pytest --nbmake .
 
   coverage:
     runs-on: ubuntu-latest
