### 2023-10-24 42b1887
ci: install package before running tests
