+++ .github/workflows/tests.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install .
           pip install pytest nbmake
       - name: Run test 
         run: |
@@ -47,7 +47,8 @@ jobs:
       - name: Install dependencies
         shell: bash -l {0}
         run: |
-          pip install -r requirements.txt
+          python -m pip install --upgrade pip
+          pip install .
           pip install pytest nbmake
       - name: Run test 
         shell: bash -l {0}
