### 2023-10-24 3a4954a
docs: add notebook showing speedup with subsampling
