### 2023-10-24 #5791 @LuisScoccola
Hi @raphaelreinauer, thank you again for your detailed feedback.

Wrt the error in the notebook, we have avoided using ipympl, and [we are also now running the notebooks in CI](https://github.com/scikit-tda/DREiMac/pull/13/commits/9adccdcfcf4bd297c63d8bcdaeb416875ab362c9), so that pytest fails if there is any issue with the notebooks. We are also catching and printing the error that we wanted to showcase in the notebook parameter_standard_range, so that tests don't fail.

Wrt to benchmarks (L40-L42 of the pdf), we have added a notebook to the repository and documentation showing this (parameter_n_samples).

Please let me know if I have forgotten anything.
