### 2023-11-04 #5791 @raphaelreinauer
Comment: Line 63 seems to be a typo: `cohomological dimensions 1, 2, and 1, respectively`
