### 2023-11-04 #5791 @raphaelreinauer
When I randomly change the order of the images in the `col20.ipynb` notebook like this:
```
#Randomly change the order of the images
np.random.seed(0)
perm = np.random.permutation(len(X))
X = X[perm]
img = [imgs[i] for i in perm]
```
the visualization of the circular map changes drastically:
![image](https://github.com/openjournals/joss-reviews/assets/16176161/a4ef1f53-d26f-4d1b-ba5a-e0326d07a9f7)
The circular coordinates shouldn't depend on the order of the coordinates. Please clarify what I misunderstood.
