### 2023-11-04 #5791 @raphaelreinauer
`K` is not defined:
https://github.com/scikit-tda/DREiMac/blob/9ced70db1f275c9fcaa78c1788477e764b1106f9/dreimac/emcoords.py#L103
