### 2023-11-04 #5791 @raphaelreinauer
@fabian-s I've completed the checklist for the review. The paper is well-written and contains all the main contributions of the software. There are minor issues that the authors should address or have already addressed; however, these are not critical for the paper's acceptance.
The public API documentation is thorough, although it relies solely on type hints within the docstrings. This approach may limit static type checking by conventional Python type checkers. The documentation for the internal code, which is not directly interacted with by the users, is less detailed, making it somewhat challenging to follow.
The examples in the notebooks are particularly useful in illustrating the capabilities of the software, especially the example utilizing the COIL-20 dataset.

In conclusion, I suggest the paper be accepted.
