### 2023-11-06 #5791 @fabian-s
Thank you @raphaelreinauer and @mtsch for  reviewing this software for JOSS.

@LuisScoccola, at this point could you please:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare or OSF.io)
