### 2023-11-15 5d63548
docs: add clarification in docstring
