### 2023-11-15 #5791 @LuisScoccola
Thank you @fabian-s, @mtsch, and @raphaelreinauer!

- The latest version tag (which I just released) is **0.3.1**, and can be found here: https://github.com/scikit-tda/DREiMac/releases/tag/0.3.1
- The DOI for Zenodo is **10.5281/zenodo.10136542**, and can be found here: https://zenodo.org/records/10136542

Please let me know if anything else is required at this point.
