### 2023-11-16 #5791 @fabian-s
note: "missing DOI" for Dionysius above is a false positive
