### 2023-11-16 #5791 @fabian-s
"missing DOI" for Dionysius above is a false positive
