### 2023-11-16 #5791 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
