### 2023-11-16 74acb7b
minor case fixes
