+++ paper/paper.bib
@@ -27,7 +27,7 @@ MRREVIEWER = {\.{I}smet\ Karaca},
 @InProceedings{toroidal-coords,
   author =	{Scoccola, Luis and Gakhar, Hitesh and Bush, Johnathan and Schonsheck, Nikolas and Rask, Tatum and Zhou, Ling and Perea, Jose A.},
   title =	{{Toroidal Coordinates: Decorrelating Circular Coordinates with Lattice Reduction}},
-  booktitle =	{39th International Symposium on Computational Geometry (SoCG 2023)},
+  booktitle =	{39th {I}nternational {S}ymposium on {C}omputational {G}eometry (SoCG 2023)},
   pages =	{57:1--57:20},
   series =	{Leibniz International Proceedings in Informatics (LIPIcs)},
   ISBN =	{978-3-95977-273-0},
@@ -82,7 +82,7 @@ MRREVIEWER = {Ne\v{z}a\ Mramor Kosta},
   editor       = {Zachary Friggstad and
                   Jean{-}Lou De Carufel},
   title        = {Coordinatizing Data With Lens Spaces and Persistent Cohomology},
-  booktitle    = {Proceedings of the 31st Canadian Conference on Computational Geometry,
+  booktitle    = {Proceedings of the 31st {C}anadian {C}onference on {C}omputational {G}eometry,
                   {CCCG} 2019, {A}ugust 8-10, 2019, {U}niversity of {A}lberta, {E}dmonton, {A}lberta,
                   {C}anada},
   pages        = {49--58},
@@ -133,7 +133,7 @@ MRREVIEWER = {Patrizio\ Frosini},
 @InProceedings{fibered,
   author =	{Scoccola, Luis and Perea, Jose A.},
   title =	{{FibeRed: Fiberwise Dimensionality Reduction of Topologically Complex Data with Vector Bundles}},
-  booktitle =	{39th International Symposium on Computational Geometry (SoCG 2023)},
+  booktitle =	{39th {I}nternational {S}ymposium on {C}omputational {G}eometry (SoCG 2023)},
   pages =	{56:1--56:18},
   series =	{Leibniz International Proceedings in Informatics (LIPIcs)},
   ISBN =	{978-3-95977-273-0},
@@ -193,7 +193,7 @@ MRREVIEWER = {Patrizio\ Frosini},
   number = {54},
   pages = {2614},
   author = {Matija Čufar},
-  title = {Ripserer.jl: flexible and efficient persistent homology computation in Julia},
+  title = {Ripserer.jl: flexible and efficient persistent homology computation in {J}ulia},
   journal = {Journal of Open Source Software}
 }
 
