### 2023-11-16 #5791 @danielskatz
I've suggested some minor changes in https://github.com/scikit-tda/DREiMac/pull/15 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
