### 2023-11-17 #5791 @danielskatz
Congratulations to @LuisScoccola (Luis Scoccola) and co-authors on your publication!!

JOSS depends on volunteers and couldn't be successful without you
